Basic multiplatform build
