+++ certifikit-cli/src/jvmMain/kotlin/app/cash/certifikit/cli/client.kt
@@ -19,9 +19,7 @@ import java.util.concurrent.TimeUnit
 import okhttp3.ConnectionSpec
 import okhttp3.OkHttpClient
 import okhttp3.tls.HandshakeCertificates
-import okio.ExperimentalFileSystem
 
-@OptIn(ExperimentalFileSystem::class)
 fun Main.buildClient(): OkHttpClient {
   return OkHttpClient.Builder()
       .connectTimeout(2, TimeUnit.SECONDS)
