+++ certifikit-cli/src/jvmMain/kotlin/app/cash/certifikit/cli/errors/errorCategorisation.kt
