+++ certifikit-cli/src/jvmMain/kotlin/app/cash/certifikit/cli/keystore.kt
@@ -18,11 +18,9 @@ package app.cash.certifikit.cli
 import java.security.KeyStore
 import javax.net.ssl.TrustManagerFactory
 import javax.net.ssl.X509TrustManager
-import okio.ExperimentalFileSystem
 import okio.FileSystem
 import okio.Path
 
-@OptIn(ExperimentalFileSystem::class)
 fun Path.trustManager(filesystem: FileSystem = FileSystem.SYSTEM): X509TrustManager {
     val factory = TrustManagerFactory.getInstance(
             TrustManagerFactory.getDefaultAlgorithm())
