+++ certifikit-cli/src/jvmTest/kotlin/app/cash/certifikit/cli/CertificateOutputTest.kt
