+++ certifikit-cli/src/jvmTest/kotlin/app/cash/certifikit/cli/TestCerts.kt
@@ -13,34 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.certifikit.text
+package app.cash.certifikit.cli
 
-import java.security.KeyPair
-import java.security.cert.X509Certificate
-import okio.ExperimentalFileSystem
 import okio.FileSystem
-import okio.Path
 import okio.Path.Companion.toPath
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-@OptIn(ExperimentalFileSystem::class)
-class PemTest {
+class TestCerts {
   val fileSystem = FileSystem.SYSTEM
 
   @Test
-  fun parseHeldCertificate() {
-    val (pkcs8pair, cert1) = decode("src/test/resources/pkcs8pair.pem".toPath())
-    val (pkcs1pair, cert2) = decode("src/test/resources/pkcs1pair.pem".toPath())
-
-    assertThat(cert1).isEqualTo(cert2)
-    assertThat(pkcs1pair.private).isEqualTo(pkcs8pair.private)
-    assertThat(pkcs1pair.public).isEqualTo(pkcs8pair.public)
-  }
-
-  private fun decode(file: Path): Pair<KeyPair, X509Certificate> {
-    return fileSystem.read(file) {
-      decode(readUtf8())
+  fun parseCert() {
+    val cert = fileSystem.read("src/test/resources/cert.pem".toPath()) {
+      readUtf8().parsePemCertificate()
     }
+    assertThat(cert.signatureAlgorithm.algorithm).isEqualTo("1.2.840.113549.1.1.11")
   }
 }
