+++ certifikit/src/jvmMain/kotlin/app/cash/certifikit/AnyValue.kt
