+++ certifikit/src/jvmMain/kotlin/app/cash/certifikit/BasicDerAdapter.kt
